[terminal] Improvements and refactoring
