+++ .scripts/packages.list
@@ -57,16 +57,17 @@ L + netutil
 L - system/process
 * - system/procname
 L + system/sensors
-* - terminal
+* + terminal
 * ! terminal/tty
+* - terminal/input
 * - terminal/window
 * + timeutil
 * + tmp
 * + usage
-* + usage/update
-* + usage/man
 * + usage/completion/bash
 * + usage/completion/fish
 * + usage/completion/zsh
+* + usage/man
+* + usage/update
 * + uuid
 * + version
