+++ terminal/terminal.go
@@ -1,6 +1,3 @@
-//go:build !windows
-// +build !windows
-
 // Package terminal provides methods for working with user input
 package terminal
 
@@ -15,43 +12,14 @@ import (
 	"fmt"
 	"os"
 	"strings"
-	"unicode/utf8"
 
-	"github.com/essentialkaos/ek/v12/ansi"
 	"github.com/essentialkaos/ek/v12/fmtc"
-	"github.com/essentialkaos/ek/v12/fsutil"
-	"github.com/essentialkaos/ek/v12/mathutil"
-	"github.com/essentialkaos/ek/v12/secstr"
-
-	"github.com/essentialkaos/go-linenoise/v3"
+	"github.com/essentialkaos/ek/v12/strutil"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// ErrKillSignal is error type when user cancel input
-var ErrKillSignal = linenoise.ErrKillSignal
-
-// Prompt is prompt string
-var Prompt = "> "
-
-// MaskSymbol is symbol used for masking passwords
-var MaskSymbol = "*"
-
-// HideLength is flag for hiding password length
-var HideLength = false
-
-// HidePassword is flag for hiding password while typing
-// Because of using the low-level linenoise method for this feature, we can not use a
-// custom masking symbol, so it always will be an asterisk (*).
-var HidePassword = false
-
 var (
-	// MaskSymbolColorTag is fmtc color tag used for MaskSymbol output
-	MaskSymbolColorTag = ""
-
-	// TitleColorTag is fmtc color tag used for input titles
-	TitleColorTag = "{s}"
-
 	// ErrorColorTag is fmtc color tag used for error messages
 	ErrorColorTag = "{r}"
 
@@ -73,78 +41,7 @@ var (
 	InfoPrefix = ""
 )
 
-// AlwaysYes is a flag, if set ReadAnswer will always return true (useful for working
-// with option for forced actions)
-var AlwaysYes = false
-
 // ////////////////////////////////////////////////////////////////////////////////// //
-
-var tmux int8
-
-// ////////////////////////////////////////////////////////////////////////////////// //
-
-// Read reads user's input
-func Read(title string, nonEmpty bool) (string, error) {
-	return readUserInput(title, nonEmpty, false)
-}
-
-// ReadAnswer reads user's answer for yes/no question
-func ReadAnswer(title string, defaultAnswers ...string) (bool, error) {
-	var defaultAnswer string
-
-	if len(defaultAnswers) != 0 {
-		defaultAnswer = defaultAnswers[0]
-	}
-
-	if AlwaysYes {
-		if title != "" {
-			fmtc.Println(TitleColorTag + getAnswerTitle(title, defaultAnswer) + "{!}")
-		}
-		fmtc.Println(Prompt + "y")
-		return true, nil
-	}
-
-	for {
-		answer, err := readUserInput(
-			getAnswerTitle(title, defaultAnswer), false, false,
-		)
-
-		if err != nil {
-			return false, err
-		}
-
-		if answer == "" {
-			answer = defaultAnswer
-		}
-
-		switch strings.ToUpper(answer) {
-		case "Y":
-			return true, nil
-		case "N":
-			return false, nil
-		default:
-			PrintWarnMessage("\nPlease enter Y or N\n")
-		}
-	}
-}
-
-// ReadPassword reads password or some private input which will be hidden
-// after pressing Enter
-func ReadPassword(title string, nonEmpty bool) (string, error) {
-	return readUserInput(title, nonEmpty, true)
-}
-
-// ReadPasswordSecure reads password or some private input which will be hidden
-// after pressing Enter
-func ReadPasswordSecure(title string, nonEmpty bool) (*secstr.String, error) {
-	password, err := readUserInput(title, nonEmpty, true)
-
-	if err != nil {
-		return nil, err
-	}
-
-	return secstr.NewSecureString(&password)
-}
 
 // PrintActionMessage prints message about action currently in progress
 func PrintActionMessage(message string) {
@@ -158,194 +55,62 @@ func PrintActionStatus(status int) {
 		fmtc.Println("{g}OK{!}")
 	case 1:
 		fmtc.Println("{r}ERROR{!}")
+	case 2:
+		fmtc.Println("{y}WARNING{!}")
+	default:
+		fmtc.Println("{s}UNKNOWN{!}")
 	}
 }
 
 // Error prints error message
 func Error(message any, args ...any) {
-	msg := formatMessage(message)
-
-	if len(args) == 0 {
-		fmtc.Fprintf(os.Stderr, ErrorPrefix+ErrorColorTag+"%s{!}\n", msg)
-	} else {
-		fmtc.Fprintf(os.Stderr, ErrorPrefix+ErrorColorTag+"%s{!}\n", fmt.Sprintf(msg, args...))
-	}
+	fmtc.Fprintf(
+		os.Stdout, ErrorColorTag+ErrorPrefix+"%s{!}\n",
+		formatMessage(message, ErrorPrefix, args),
+	)
 }
 
 // Warn prints warning message
 func Warn(message any, args ...any) {
-	msg := formatMessage(message)
-
-	if len(args) == 0 {
-		fmtc.Fprintf(os.Stderr, WarnPrefix+WarnColorTag+"%s{!}\n", msg)
-	} else {
-		fmtc.Fprintf(os.Stderr, WarnPrefix+WarnColorTag+"%s{!}\n", fmt.Sprintf(msg, args...))
-	}
+	fmtc.Fprintf(
+		os.Stdout, WarnColorTag+WarnPrefix+"%s{!}\n",
+		formatMessage(message, WarnPrefix, args),
+	)
 }
 
 // Info prints info message
 func Info(message any, args ...any) {
-	msg := formatMessage(message)
-
-	if len(args) == 0 {
-		fmtc.Fprintf(os.Stdout, InfoPrefix+InfoColorTag+"%s{!}\n", msg)
-	} else {
-		fmtc.Fprintf(os.Stdout, InfoPrefix+InfoColorTag+"%s{!}\n", fmt.Sprintf(msg, args...))
-	}
-}
-
-// AddHistory adds line to input history
-func AddHistory(data string) {
-	linenoise.AddHistory(data)
-}
-
-// SetCompletionHandler adds function for autocompletion
-func SetCompletionHandler(h func(input string) []string) {
-	linenoise.SetCompletionHandler(h)
-}
-
-// SetHintHandler adds function for input hints
-func SetHintHandler(h func(input string) string) {
-	linenoise.SetHintHandler(h)
-}
-
-// DEPRECATED /////////////////////////////////////////////////////////////////////// //
-
-// ReadUI reads user's input
-//
-// Deprecated: Use method Read instead
-func ReadUI(title string, nonEmpty bool) (string, error) {
-	return Read(title, nonEmpty)
-}
-
-// PrintErrorMessage prints error message
-//
-// Deprecated: Use method Error instead
-func PrintErrorMessage(message string, args ...any) {
-	Error(message, args...)
-}
-
-// PrintWarnMessage prints warning message
-//
-// Deprecated: Use method Warn instead
-func PrintWarnMessage(message string, args ...any) {
-	Warn(message, args...)
+	fmtc.Fprintf(
+		os.Stdout, InfoColorTag+InfoPrefix+"%s{!}\n",
+		formatMessage(message, InfoPrefix, args),
+	)
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // formatMessage formats message based on it type
-func formatMessage(message any) string {
+func formatMessage(message any, prefix string, args []any) string {
+	var msg string
+
 	switch m := message.(type) {
 	case string:
-		return m
+		msg = m
 	case error:
-		return m.Error()
-	}
-
-	return fmt.Sprint(message)
-}
-
-// getMask returns mask for password
-func getMask(message string) string {
-	var masking string
-
-	// Remove fmtc color tags and ANSI escape codes
-	prompt := fmtc.Clean(ansi.RemoveCodes(Prompt))
-	prefix := strings.Repeat(" ", utf8.RuneCountInString(prompt))
-	length := utf8.RuneCountInString(message)
-
-	if !HideLength {
-		if isTmuxSession() {
-			masking = strings.Repeat("*", length)
-		} else {
-			masking = strings.Repeat(MaskSymbol, length)
-		}
-	} else {
-		masking = "[hidden]" + strings.Repeat(" ", mathutil.Max(0, length-8))
-	}
-
-	if !fsutil.IsCharacterDevice("/dev/stdin") && os.Getenv("FAKETTY") == "" {
-		return fmtc.Sprintf(Prompt) + masking
-	}
-
-	return fmt.Sprintf("%s\033[1A%s", prefix, masking)
-}
-
-// getAnswerTitle returns title with info about default answer
-func getAnswerTitle(title, defaultAnswer string) string {
-	if title == "" {
-		return ""
-	}
-
-	switch strings.ToUpper(defaultAnswer) {
-	case "Y":
-		return fmt.Sprintf(TitleColorTag+"%s ({*}Y{!*}/n){!}", title)
-	case "N":
-		return fmt.Sprintf(TitleColorTag+"%s (y/{*}N{!*}){!}", title)
+		msg = m.Error()
 	default:
-		return fmt.Sprintf(TitleColorTag+"%s (y/n){!}", title)
-	}
-}
-
-// readUserInput reads user input
-func readUserInput(title string, nonEmpty, private bool) (string, error) {
-	if title != "" {
-		fmtc.Println(TitleColorTag + title + "{!}")
+		msg = fmt.Sprint(message)
 	}
 
-	var input string
-	var err error
-
-	if private && HidePassword {
-		linenoise.SetMaskMode(true)
+	if len(args) > 0 {
+		msg = fmt.Sprintf(msg, args...)
 	}
 
-	for {
-		input, err = linenoise.Line(fmtc.Sprintf(Prompt))
-
-		if private && HidePassword {
-			linenoise.SetMaskMode(false)
-		}
-
-		if err != nil {
-			return "", err
-		}
-
-		if nonEmpty && strings.TrimSpace(input) == "" {
-			PrintWarnMessage("\nYou must enter non-empty value\n")
-			continue
-		}
-
-		if private && input != "" {
-			if !HidePassword {
-				if MaskSymbolColorTag == "" {
-					fmt.Println(getMask(input))
-				} else {
-					fmtc.Println(MaskSymbolColorTag + getMask(input) + "{!}")
-				}
-			}
-		} else {
-			if !fsutil.IsCharacterDevice("/dev/stdin") && os.Getenv("FAKETTY") == "" {
-				fmt.Println(fmtc.Sprintf(Prompt) + input)
-			}
-		}
-
-		break
+	if prefix != "" && strings.Contains(msg, "\n") {
+		prefixOffset := strings.Repeat(" ", strutil.Len(prefix))
+		msgText := strings.TrimRight(msg, "\n")
+		tail := strutil.Substr(msg, len(msgText), 999999)
+		msg = strings.ReplaceAll(msgText, "\n", "\n"+prefixOffset) + tail
 	}
 
-	return input, err
-}
-
-// isTmuxSession returns true if we work in tmux session
-func isTmuxSession() bool {
-	if tmux == 0 {
-		if os.Getenv("TMUX") == "" {
-			tmux = -1
-		} else {
-			tmux = 1
-		}
-	}
-
-	return tmux == 1
+	return msg
 }
